requirements.txt
networkx==2.1
everett==0.9
pluginbase==0.7
npstreams==1.4.*
lxml==4.2.*
shapely==1.6.*
pyproj==1.9.5.*
imageio==2.3.*
scikit-learn==0.19.*
six==1.11.*
h5py==2.7.*
matplotlib==2.1.*
pillow==5.0.*
click==6.*
rasterio==1.0.7
protobuf==3.6.*
scipy==1.1.*
